refactor: Rename "Common" to "common".
