+++ src/Questions/index.js
@@ -2,10 +2,10 @@ import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 import ImageZoom from 'react-medium-image-zoom';
 
-import countries from '../Common/countries';
+import countries from '../common/countries';
 import insightTypes from './insightTypes';
 import './questions.css';
-import subDomain from '../Common/subdomain';
+import subDomain from '../common/subdomain';
 
 const NO_QUESTION_REMAINING = 'NO_QUESTION_REMAINING';
 
@@ -60,12 +60,18 @@ const Questions = () => {
           )
           .map(q => ({
             ...q,
-            productLink: `https://${subDomain.subDomain}.openfoodfacts.org/product/${q.barcode}`,
+            productLink: `https://${
+              subDomain.subDomain
+            }.openfoodfacts.org/product/${q.barcode}`,
           }));
         return axios.all(
           questionsResults.map(q =>
             axios(
-              `https://${subDomain.subDomain}.openfoodfacts.org/api/v0/product/${q.barcode}.json?fields=product_name`,
+              `https://${
+                subDomain.subDomain
+              }.openfoodfacts.org/api/v0/product/${
+                q.barcode
+              }.json?fields=product_name`,
             ),
           ),
         );
@@ -89,7 +95,9 @@ const Questions = () => {
     axios.post(
       'https://robotoff.openfoodfacts.org/api/v1/insights/annotate',
       new URLSearchParams(
-        `insight_id=${questions[0].insight_id}&annotation=${annotation}&update=1`,
+        `insight_id=${
+          questions[0].insight_id
+        }&annotation=${annotation}&update=1`,
       ),
       { withCredentials: true },
     ); // The status of the response is not displayed so no need to wait the response
